+++ packages/app/src/account/query/use-cached-balance.query.ts
@@ -2,13 +2,18 @@ import { useLiveQuery } from 'drizzle-orm/expo-sqlite';
 
 import { accountBalanceRepository } from '../../@generic/drizzle/db/db';
 
+import { useAccountBalancesUpdatedAtQuery } from './use-account-balances-updated-at.query';
 import { useCachedMicroUnitQuery } from './use-cached-micro-unit.query';
 
-type BalanceQuery = ReturnType<typeof accountBalanceRepository.getByAccountId>;
+type BalanceQuery =
+    | ReturnType<typeof accountBalanceRepository.getByAccountId>
+    | ReturnType<typeof accountBalanceRepository.getArchivedAccountBalance>;
 
 export const useCachedBalanceQuery = (query: BalanceQuery, dependencies: unknown[]) => {
-    const { data } = useLiveQuery(query, dependencies);
-    const balance = useCachedMicroUnitQuery(data.at(0)?.balance, dependencies);
+    const accountBalancesUpdatedAt = useAccountBalancesUpdatedAtQuery();
+    const queryDependencies = [...dependencies, accountBalancesUpdatedAt];
+    const { data } = useLiveQuery(query, queryDependencies);
+    const balance = useCachedMicroUnitQuery(data.at(0)?.balance, queryDependencies);
 
     return { balance };
 };