+++ packages/app/src/transaction/components/split-entries-add-item-footer/split-entries-add-item-footer.selector.ts
@@ -1,0 +1,3 @@
+export const SplitEntriesAddItemFooterSelector = {
+    AddItemButton: 'SplitEntries.AddItemButton'
+} as const;